✨ feat(validation): Implementação de validação com zod em diversos controllers de notícias e equipes, garantindo a integridade dos dados recebidos.

+++ src/controllers/news/delete-news-controller.ts
@@ -1,7 +1,13 @@
 import type { Request, Response } from 'express'
+import z from 'zod'
 import { HttpStatus } from '../../@types/status-code.ts'
 import type { INewsRepository } from '../../repositories/news/inews-repository.d.ts'
-import type { IFirebaseStorageService } from '../../services/firebase-storage/ifirebase-storage.js'
+import type { IFirebaseStorageService } from '../../services/firebase-storage/ifirebase-storage.ts'
+
+const deleteNewsSchema = z.object({
+  id: z.string(),
+})
+
 export class DeleteNewsController {
   constructor(
     private readonly newsRepository: INewsRepository,
@@ -10,7 +16,15 @@ export class DeleteNewsController {
 
   async handle(req: Request, res: Response) {
     try {
-      const { id } = req.params
+      const parseResult = deleteNewsSchema.safeParse(req.params)
+
+      if (!parseResult.success) {
+        return res.status(HttpStatus.BAD_REQUEST).json({
+          errors: z.treeifyError(parseResult.error),
+        })
+      }
+
+      const { id } = parseResult.data
 
       await Promise.all([
         this.newsRepository.delete(id),
